style: tightened up trending content for smaller screens
