+++ frontend/src/Components/dashboard/index.ts
@@ -1,2 +1,2 @@
 export { default as ResidentWeeklyActivityTable } from './ResidentWeeklyActivityTable';
-export { default as FeaturedContent } from './FeaturedContent';
+export { default as ExpandableCardGrid } from './ExpandableCardGrid';
