+++ frontend/src/Pages/AdminLayer1.tsx
@@ -2,11 +2,13 @@ import {
     FilterPastTime,
     Library,
     OpenContentItem,
-    ServerResponseMany
+    ServerResponseMany,
+    UserRole
 } from '@/common';
-import { FeaturedContent } from '@/Components/dashboard';
 import TopContentList from '@/Components/dashboard/TopContentList';
 import DropdownControl from '@/Components/inputs/DropdownControl';
+import { ExpandableCardGrid } from '@/Components/dashboard';
+import LibraryCard from '@/Components/LibraryCard';
 import { AxiosError } from 'axios';
 import { useEffect, useState } from 'react';
 import { useLoaderData, useNavigate } from 'react-router-dom';
@@ -34,7 +36,21 @@ export default function AdminLayer1() {
     }, [timeFilter]);
     return (
         <div className="w-full flex flex-col gap-6 px-5 pb-4">
-            <FeaturedContent featured={featured} />
+            <ExpandableCardGrid
+                items={featured}
+                title="Featured Content"
+                emptyStateLink="/knowledge-center-management/libraries"
+                emptyStateText="Feature content to showcase"
+            >
+                {(item) => (
+                    <LibraryCard
+                        library={item}
+                        key={item.id}
+                        role={UserRole.Student}
+                    />
+                )}
+            </ExpandableCardGrid>
+
             <div className="flex flex-row justify-between items-center">
                 <h2>Insights</h2>
                 <DropdownControl
