+++ frontend/src/Pages/StudentLayer1.tsx
@@ -3,17 +3,19 @@ import {
     OpenContentItem,
     UserRole,
     ServerResponseOne,
-    HelpfulLink,
     HelpfulLinkAndSort,
     ServerResponseMany
 } from '@/common';
 import HelpfulLinkCard from '@/Components/cards/HelpfulLinkCard';
+import LibraryCard from '@/Components/LibraryCard';
 import { useLoaderData, useNavigate } from 'react-router-dom';
-import { FeaturedContent } from '@/Components/dashboard';
+import { ExpandableCardGrid } from '@/Components/dashboard';
 import TopContentList from '@/Components/dashboard/TopContentList';
 import useSWR from 'swr';
 import { AxiosError } from 'axios';
 import OpenContentItemAccordion from '@/Components/OpenContentItemAccordion';
+import { useEffect, useRef, useState } from 'react';
+import LibrarySearchResultsModal from '@/Components/LibrarySearchResultsModal';
 
 export default function StudentLayer1() {
     const navigate = useNavigate();
@@ -33,6 +35,30 @@ export default function StudentLayer1() {
         ServerResponseOne<HelpfulLinkAndSort>,
         AxiosError
     >(`api/helpful-links`);
+    const [searchModalLibrary, setSearchModalLibrary] =
+        useState<Library | null>(null);
+    const modalRef = useRef<HTMLDialogElement>(null);
+    useEffect(() => {
+        if (searchModalLibrary && modalRef.current) {
+            modalRef.current.style.visibility = 'visible';
+            modalRef.current.showModal();
+        }
+    }, [searchModalLibrary]);
+    const openSearchModal = (library: Library) => {
+        setSearchModalLibrary(library); //fire off useEffect
+    };
+    const closeSearchModal = () => {
+        if (modalRef.current) {
+            modalRef.current.style.visibility = 'hidden';
+            modalRef.current.close();
+        }
+        setSearchModalLibrary(null);
+    };
+    const navToLibraryViewer = (url: string, title: string) => {
+        navigate(`/viewer/libraries/${searchModalLibrary?.id}`, {
+            state: { url: url, title: title }
+        });
+    };
 
     function navigateToOpenContent() {
         navigate('/knowledge-center/libraries');
@@ -46,12 +72,32 @@ export default function StudentLayer1() {
 
     return (
         <div className="flex flex-row h-full">
+            {searchModalLibrary && (
+                <LibrarySearchResultsModal
+                    ref={modalRef}
+                    libraryId={searchModalLibrary.id}
+                    searchPlaceholder={`Search ${searchModalLibrary.title}`}
+                    onItemClick={navToLibraryViewer}
+                    onModalClose={closeSearchModal}
+                    useInternalSearchBar={true}
+                />
+            )}
             {/* main section */}
             <div className="w-full flex flex-col gap-6 px-5 pb-4">
-                <FeaturedContent
-                    featured={featured?.data ?? []}
-                    mutate={updateFavorites}
-                />
+                <ExpandableCardGrid
+                    items={featured?.data ?? []}
+                    title="Featured Content"
+                >
+                    {(item) => (
+                        <LibraryCard
+                            key={item.id}
+                            library={item}
+                            role={UserRole.Student}
+                            mutate={updateFavorites}
+                            onSearchClick={() => openSearchModal(item)}
+                        />
+                    )}
+                </ExpandableCardGrid>
                 <h2> Pick Up Where You Left Off</h2>
                 <div className="grid grid-cols-2 gap-6">
                     <TopContentList
@@ -65,23 +111,24 @@ export default function StudentLayer1() {
                         navigateToOpenContent={navigateToOpenContent}
                     />
                 </div>
-                <h2>Helpful Links</h2>
-                <div
-                    className={`card card-row-padding grid grid-cols-${helpfulLinks?.data.helpful_links.length} gap-3`}
+
+                <ExpandableCardGrid
+                    items={helpfulLinks?.data.helpful_links ?? []}
+                    emptyStateText="Add helpful links to share"
+                    emptyStateLink="knowledge-center/helpful-links"
                 >
-                    {helpfulLinks?.data.helpful_links.map(
-                        (link: HelpfulLink) => (
-                            <HelpfulLinkCard
-                                link={link}
-                                role={UserRole.Student}
-                                mutate={updateFavorites}
-                            />
-                        )
+                    {(link) => (
+                        <HelpfulLinkCard
+                            key={link.id}
+                            link={link}
+                            role={UserRole.Student}
+                            mutate={updateFavorites}
+                        />
                     )}
-                </div>
+                </ExpandableCardGrid>
             </div>
             {/* right sidebar */}
-            <div className="min-w-[390px] border-l border-grey-1 flex flex-col gap-6 px-6 py-4">
+            <div className="min-w-[290px] xl:min-w-[390px] border-l border-grey-1 flex flex-col gap-6 px-6 py-4">
                 <h2>Favorites</h2>
                 <div className="space-y-3 w-full">
                     {favorites?.data && favorites.data.length > 0 ? (
